Initialize pdf-tools without interaction from user
Applies the first time it is run, and after updates. Also will no longer
fail for not finding epdfinfo the first time it is run, and instead will
probably fail if the user tries to execute epdfinfo before restarting
emacs the first time it is run.

Also removed the package-refresh line, which should have happened in the
last commit.